IGNITE-19207 / move the configuration to catalog module
