+++ modules/catalog/build.gradle
@@ -20,6 +20,7 @@ apply from: "$rootDir/buildscripts/publishing.gradle"
 apply from: "$rootDir/buildscripts/java-junit5.gradle"
 
 dependencies {
+    annotationProcessor project(':ignite-configuration-annotation-processor')
     annotationProcessor libs.auto.service
 
     implementation project(':ignite-api')
@@ -27,6 +28,7 @@ dependencies {
     implementation project(':ignite-configuration')
     implementation project(':ignite-metastorage-api')
     implementation libs.jetbrains.annotations
+    implementation libs.auto.service.annotations
 
     testImplementation(testFixtures(project(':ignite-core')))
     testImplementation libs.mockito.junit
