+++ modules/catalog/src/main/java/org/apache/ignite/internal/catalog/configuration/SchemaSynchronizationConfigurationModule.java
@@ -15,26 +15,27 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.internal.schemasync.configuration;
+package org.apache.ignite.internal.catalog.configuration;
 
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.contains;
-import static org.hamcrest.Matchers.is;
-
+import com.google.auto.service.AutoService;
+import java.util.Collection;
+import java.util.Set;
 import org.apache.ignite.configuration.ConfigurationModule;
+import org.apache.ignite.configuration.RootKey;
 import org.apache.ignite.configuration.annotation.ConfigurationType;
-import org.junit.jupiter.api.Test;
 
-class SchemaSynchronizationConfigurationModuleTest {
-    private final ConfigurationModule module = new SchemaSynchronizationConfigurationModule();
-
-    @Test
-    void typeIsDistributed() {
-        assertThat(module.type(), is(ConfigurationType.DISTRIBUTED));
+/**
+ * {@link ConfigurationModule} for distributed (cluster-wide) configuration of Schema Synchronization.
+ */
+@AutoService(ConfigurationModule.class)
+public class SchemaSynchronizationConfigurationModule implements ConfigurationModule {
+    @Override
+    public ConfigurationType type() {
+        return ConfigurationType.DISTRIBUTED;
     }
 
-    @Test
-    void rootKeysAreAsExpected() {
-        assertThat(module.rootKeys(), contains(SchemaSynchronizationConfiguration.KEY));
+    @Override
+    public Collection<RootKey<?, ?>> rootKeys() {
+        return Set.of(SchemaSynchronizationConfiguration.KEY);
     }
 }
