--- modules/schema-sync/README.md
@@ -1,4 +1,0 @@
-# Schema Synchronization module
-
-This module provides functionality for [Schema Synchronization]
-(https://cwiki.apache.org/confluence/display/IGNITE/IEP-98%3A+Schema+Synchronization).
