+++ settings.gradle
@@ -31,7 +31,6 @@ include(':ignite-configuration-api')
 include(':ignite-raft-api')
 include(':ignite-raft')
 include(':ignite-schema')
-include(':ignite-schema-sync')
 include(':ignite-storage-page-memory')
 include(':ignite-transactions')
 include(':ignite-client-common')
@@ -86,7 +85,6 @@ project(":ignite-raft-api").projectDir = file('modules/raft-api')
 project(":ignite-configuration-api").projectDir = file('modules/configuration-api')
 project(":ignite-raft").projectDir = file('modules/raft')
 project(":ignite-schema").projectDir = file('modules/schema')
-project(":ignite-schema-sync").projectDir = file('modules/schema-sync')
 project(":ignite-storage-page-memory").projectDir = file('modules/storage-page-memory')
 project(":ignite-transactions").projectDir = file('modules/transactions')
 project(":ignite-client-common").projectDir = file('modules/client-common')
